+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.4.1]
 
 ### Fixes
+- XMLMunger can now cope with unusual namespaces better: this fixes an error when streaming entities from CAFE (and possibly others)
 - Entity metadata IDs (rather then Entity IDs) have been fixed
 - Non-compliant validUntil dates that lack a timezone offset will be processed as UTC now with a warning rather than an exception.
 - Tags should now be unique (previously passing atoms and strings could lead to duplicates)